+++ monefystat_api/app.py
@@ -1,12 +1,15 @@
 from os.path import dirname, abspath
+from json import dumps
+
 from sanic import Sanic
 from sanic.response import json
 from sanic.response import text
+from sanic.response import stream
 from sanic.request import RequestParameters
 
 from ..keys import DROPBOX_TOKEN
 from ..transport.data_provider import DataProvider
-from service_resource import drop_db_endpoint, create_db_endpoint
+from .service_resource import drop_db_endpoint, create_db_endpoint, get_all_data_endpoint
 
 PATH = dirname(dirname(abspath(__file__))) + '/downloads/Monefy_data.csv'
 DIR = '/monefy'
@@ -54,6 +57,15 @@ async def drop_endpoint(request):
         status=200
     )
 
+
+@app.get("/data")
+async def data_endpoint(request):
+    async def streaming_from_db(response):
+        result = get_all_data_endpoint()
+        for i, data in enumerate(result):
+            data = dumps(result[i])
+            response.write(data)
+    return stream(streaming_from_db, content_type='json')
 
 if __name__ == "__main__":
     app.run(host="0.0.0.0", port=80)