+++ tests/db_endpoint_test.py
@@ -21,6 +21,14 @@ class Drop_db_test(unittest.TestCase):
         request, response = app.test_client.put('/drop_db')
         assert response.status == 405
 
+class Data_db_test(unittest.TestCase):
+    def test_data_returns_200(self):
+        request, response = app.test_client.get('/data')
+        assert response.status == 200
+
+    def test_data_put_not_allowed(self):
+        request, response = app.test_client.put('/data')
+        assert response.status == 405
 
 if __name__ == "__main__":
     unittest.main()